Continuation of #1385: start refactoring test utilities
