+++ src/test/java/tools/jackson/core/fuzz/Fuzz52688ParseTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import tools.jackson.core.*;
 import tools.jackson.core.exc.StreamReadException;
 import tools.jackson.core.json.JsonFactory;
-import tools.jackson.core.testsupport.ThrottledInputStream;
+import tools.jackson.core.testutil.ThrottledInputStream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
