+++ src/test/java/tools/jackson/core/json/async/AsyncNumberCoercionTest.java
@@ -11,7 +11,7 @@ import tools.jackson.core.JsonToken;
 import tools.jackson.core.async.AsyncTestBase;
 import tools.jackson.core.exc.InputCoercionException;
 import tools.jackson.core.json.JsonFactory;
-import tools.jackson.core.testsupport.AsyncReaderWrapper;
+import tools.jackson.core.testutil.AsyncReaderWrapper;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
