+++ src/test/java/tools/jackson/core/testutil/StringWriterForTesting.java
@@ -1,4 +1,4 @@
-package tools.jackson.core.testsupport;
+package tools.jackson.core.testutil;
 
 import java.io.IOException;
 import java.io.StringWriter;
